2018-4-11

+++ appinterface/src/main/java/com/emi/roots/entity/AzForum.java
@@ -23,7 +23,7 @@ public class AzForum {
 
     private Byte state;
 
-    private Integer create_time;
+    private Long create_time;
 
     private Integer uid;
 
@@ -117,11 +117,11 @@ public class AzForum {
         this.state = state;
     }
 
-    public Integer getCreate_time() {
+    public Long getCreate_time() {
         return create_time;
     }
 
-    public void setCreate_time(Integer create_time) {
+    public void setCreate_time(Long create_time) {
         this.create_time = create_time;
     }
