+++ appinterface/src/main/java/com/emi/roots/entity/AzForumPost.java
@@ -7,7 +7,7 @@ public class AzForumPost {
 
     private Integer uid;
 
-    private Integer create_time;
+    private Long create_time;
 
     private String thumb;
 
@@ -55,11 +55,11 @@ public class AzForumPost {
         this.uid = uid;
     }
 
-    public Integer getCreate_time() {
+    public Long getCreate_time() {
         return create_time;
     }
 
-    public void setCreate_time(Integer create_time) {
+    public void setCreate_time(Long create_time) {
         this.create_time = create_time;
     }
